restart the form
