+++ src/components/index.js
@@ -1,5 +1,7 @@
 export { default as ZoeLogo } from './ZoeLogo';
 export { default as DevConsole, useDevConsole, DevConsoleProvider } from './DevConsole';
 export { default as OptionalFormattedMessage } from './OptionalFormattedMessage';
+export { default as LoanInfo } from './LoanInfo';
+export { default as PersonalInfo } from './PersonalInfo';
 
 export * from './forms';
