+++ src/intl/cs.json
@@ -6,7 +6,7 @@
 	"example.pageHome": "Domů",
 	"example.pageTerms": "Podmínky",
 	"example.pageTryIt": "Vyzkoušejte",
-	"example.personalInfoTitle": "Osobní ůdaje",
+	"example.personalInfoTitle": "Osobní údaje",
 	"example.setupLoanTitle": "Nastavte si Vaší půjčku",
 	"example.netIncomeMain": "Čistý měsíční příjem",
 	"example.givenName": "Jméno",
