+++ src/intl/en.json
@@ -21,7 +21,7 @@
 	"example.loanInfoAmount": "Select loan amount",
 	"example.totalAmountPerMonth": "Total monthly fee",
 	"example.numberOfInstalments": "Choose repayment period",
-	"example.terms": "I agreee with <a>terms and conditions</a>",
+	"example.terms": "I agree with <a>terms and conditions</a>",
 	"example.maritalStatus": "Marital status",
 	"example.maritalStatus_SINGLE": "Single",
 	"example.maritalStatus_COHABITATING": "Cohabitating",
